requirements.txt
fastapi
uvicorn
httpx
python-jose
passlib[bcrypt]
pydantic
python-jose
